virtio_scsi: RPM building
This commit edits the imported templates to build rpm binary packages
for virtio_scsi.  This process is automated with some make in
GNUmakefile, which is favored when "make" is run on the command line,
allowing us to keep this part of the build process separate from the
kbuild Makefile.

Notes:
- Dropped xen kernel building by default.
- We add the kernel "uname -r" to the package to allow side-by-side
  install of multiple driver versions for concurrent kernel version
  installs.  ABI is broken internally between RHEL5 kernels for
  virtio_confifg so there is no use trying to make a single driver build
  work for all kernel versions.

Change-Id: I4ad2cbae29122d8ade53466ab6081f9859cd8491